requirements.txt
django==5.2.5
djangorestframework==3.16.1
pytest-django==4.11.1
Faker==37.12.0